refactor(character): extract D&D business rules into CharacterRulesExt.kt
